+++ okitweb/static/okit/model/js/resources/load_balancer.js
@@ -33,7 +33,7 @@ class LoadBalancer extends OkitArtifact {
         this.merge(data);
         this.convert();
         // Expose subnet_id for the first Mount target at the top level
-        Object.defineProperty(this, 'subnet_id', {get: function() {return this.subnet_ids[0];}, set: function(subnet_id) {this.subnet_ids[0] = subnet_id;}, enumerable: false });
+        Object.defineProperty(this, 'subnet_id', {get: function() {return this.subnet_ids[0] ? this.subnet_ids[0] : '';}, set: function(subnet_id) {this.subnet_ids[0] = subnet_id;}, enumerable: false });
     }
 
     /*